testing/web-platform/tests/old-tests/submission/Infraware/Forms/contents/Forms/form_getterindex_2.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>Form_getterindex[]</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      id="input_form">
  <p><input type=hidden name="custname"></p>
  <p><input type=hidden name="custtel"></p>
  <p><input type=hidden name="custemail"></p>
 
  </form>
  <script>

    var form = document.getElementById("input_form");
    var ele;

    try
    {
      ele = form[1];
      
      test(function() {
        assert_equals(ele.name, "custtel", "getter[index] attribute is not correct.");
      });
    }
    catch (e)
    {
      test(function() {
        assert_unreached("getter[index] is not supported.");
      });
    }


  </script>

 </body>
</html>
